Rename model files to be in singular.
This is the convention in the codebase.
