+++ jobserver/models/__init__.py
@@ -14,7 +14,7 @@ from .release_file import ReleaseFile
 from .release_file_review import ReleaseFileReview
 from .repo import Repo
 from .report import Report
-from .site_alerts import SiteAlert
+from .site_alert import SiteAlert
 from .snapshot import Snapshot
 from .stats import Stats
 from .user import User
